+++ ui/src/components/core/button/button-icon.styles.ts
