+++ ui/src/components/core/button/index.ts
@@ -1,2 +1,3 @@
 export * from './button.styles';
 export * from './button';
+export * from './icon-button';
