+++ ui/src/components/core/index.ts
@@ -1,1 +1,2 @@
 export * from './button';
+export * from './icon';
