+++ ui/src/components/layout/index.ts
@@ -1,2 +1,3 @@
 export * from './grid.styles';
 export * from './flex.styles';
+export * from './stepper';
